[Refactor] 테스트 페이지 폴더 이동(test 폴더 생성 후 이동)
