+++ my-blog-app/src/pages/test/ConfirmModalTestPage.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import ConfirmModal from '../components/common/ConfirmModal/ConfirmModal'
+import ConfirmModal from '../../components/common/ConfirmModal/ConfirmModal'
 
 export default function ConfirmModalTestPage() {
   const [isOpenWithDesc, setIsOpenWithDesc] = useState(false)
